Share scalarization pass between CUDA and ROCM
This patch pulls out the scalarization pass and shares
it between CUDA and ROCM. With this change, Mobilenet
now compiles on ROCM.
