+++ iree/compiler/Conversion/LinalgToLLVMGPU/test/scalarize.mlir
@@ -1,4 +1,4 @@
-// RUN: iree-opt -iree-cuda-scalarize-math-op %s | IreeFileCheck %s
+// RUN: iree-opt -iree-llvmgpu-scalarize-math-op %s | IreeFileCheck %s
 
 // CHECK-LABEL: func @scalarize
 func @scalarize(
